# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malmö Resturant Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/itemimages/unreport/{id}": {
            "put": {
                "tags": [
                    "Item Images"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ItemImagesController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/itemreviews/unreport/{id}": {
            "put": {
                "tags": [
                    "Item Reviews"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ItemReviewsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/items/delete/{id}": {
            "delete": {
                "tags": [
                    "Items"
                ],
                "description": "Delete a specific item",
                "operationId": "App\\Http\\Controllers\\ItemsController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Delete a specific item"
                    }
                }
            }
        },
...
        "/api/items/unreport/{id}": {
            "put": {
                "tags": [
                    "Items"
                ],
                "description": "unreport an item",
                "operationId": "App\\Http\\Controllers\\ItemsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item"
                    }
                }
            }
        },
...
        "/api/resturantimages/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturant Images"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ResturantImagesController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/resturantreviews/new": {
            "post": {
                "tags": [
                    "Resturant Reviews"
                ],
                "description": "Create a new Resturant Review",
                "operationId": "App\\Http\\Controllers\\ResturantReviewsController::create",
                "parameters": [
                    {
                        "name": "reviewer",
                        "in": "query",
                        "description": "Name of Reviewer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Rating of resturant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "review",
                        "in": "query",
                        "description": "Review of Resturant",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "resturant_id",
                        "in": "query",
                        "description": "Id of resturant it belongs to.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Create a new Resturant Review"
                    }
                }
            }
        },
        "/api/resturantreviews/{id}": {
            "get": {
                "tags": [
                    "Resturant Reviews"
                ],
                "description": "Get a specific resturant review",
                "operationId": "App\\Http\\Controllers\\ResturantReviewsController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of resturant review",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a specific resturant review"
                    }
                }
            }
        },
        "/api/resturantreviews/report/{id}": {
            "put": {
                "tags": [
                    "Resturant Reviews"
                ],
                "description": "Report a Resturant Review",
                "operationId": "App\\Http\\Controllers\\ResturantReviewsController::report",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of Resturant Review",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Report an item"
                    }
                }
            }
        },
        "/api/resturantreviews/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturant Reviews"
                ],
                "description": "unreport an Resturant Review",
                "operationId": "App\\Http\\Controllers\\ResturantReviewsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an Resturant Review"
                    }
                }
            }
        },
...
        "/api/resturants/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturants"
                ],
                "description": "unreport a Resturant",
                "operationId": "App\\Http\\Controllers\\ResturantsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of resturant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport a Resturant"
                    }
                }
            }
        }
    }
}